Fix localization bug (#196)
* Improvements. See below.

Added a cookie setting and a link to the privacy policy, fixed the translation issue due to the plug-in having changed how it structures languages. Incremented the nav: section to follow proper YAML structure.

* Updated plug-ins and markdown versions

There was an issue with translations not showing up on the site anymore, for a long time. Turns out many plug-ins were using old versions.

* Fixed `../CONTRIBUTING.md` incorrect link

* Fixed 2 issues that made the script non-functional

Also added alternative sets of commands for macOS and Linux.

* Fixed multiple translation errors

Simpleen garbled up many URLs and links

* fixed error in redirection to non-existent file

* Renamed all files with `pt_BR` to `pt-BR`

This is to match the new version of the translation plug-in.

* More translations

* Added config for validation of nav and links

In order to lessen the number of info messages while building.